# 0.2.5 < RtSearch_TagLine,GoogleSocialLogin Refactorying>
#################################
1. ✨ RtSearch_TagLine
2. ♻️ GoogleSocialLogin Refactorying
3.
4.

#################################
✨ feat: 새로운 기능 추가
🐛 fix: 버그 수정
🎨 style: 코드 포맷팅, 스타일 변경 (기능 변경 없음)
♻️ refactor: 코드 리팩터링
✅ test: 테스트 코드 추가 또는 수정
🔨 chore: 빌드 관련 설정 변경 또는 보조 작업
🚀 deploy: 배포 관련 작업
🗑️ remove: 파일 또는 코드 삭제
