+++ .github/workflows/checks.yml
@@ -64,7 +64,7 @@ jobs:
         run: |
           uv run black . --check
 
-      # db 연결을 테스트ㅎ
+      # db 연결을 테스트
       - name: Wait for MySQL
         run: |
           while ! mysqladmin ping -h "127.0.0.1" -P 3306 --silent; do
