+++ app/config/tortoise_config.py
@@ -10,7 +10,7 @@ TORTOISE_APP_MODELS: list[str] = [
     "app.models.token_blacklist",
     "app.models.RSO_user",
     "app.models.refresh_token",
-    "app.models.rtSearch",
+    "app.models.search_summoner",
 
 ]
