+++ app/dtos/riot_user_dto.py
@@ -1,7 +1,6 @@
-
 from tortoise.contrib.pydantic import pydantic_model_creator
 
-from app.models.rtSearch import RtSearchModel
+from app.models.search_summoner import RtSearchModel
 
 RiotUser_Pydantic = pydantic_model_creator(RtSearchModel)
